main.py
This is a python code that uses Open Weather map API to fetch data  for my current weather and determines if it will rain. It then uses the Twilio module to send me an SMS on my phone if it's gonna rain. It also uses the Python Anywhere site to schedule time for sending me the alert message daily.